+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## [Unreleased]
+### Fixed
+- Implements stricter credit cards pattern matching
+
 ## [0.2.1] - 2016-12-19
 ### Changed
 - Updates README for usage with Rails